Move to Jetpack Compose
This replaces the main app with one rebuilt from the ground up with Jetpack Compose. The aim right now is to reach feature parity with the legacy app, which I think has been achieved.

This pretty much replaces everything in the repo with my local application. It would be a miracle if everything works first try. Code formatting has not been done. There will be random comments with old code. It'll be sorted later.

Among a visual overhaul supporting system dark theme as well as the system colour scheme on Android 12+, themed icon support, and predictive back support, this also bumps the minimum SDK version to 26 (Android 8.0). Maybe if all non-adaptive launcher icon resources are added, it might work on lower versions, I don't know. It has been tested on Android 13 and I can't guarantee compatibility with old versions yet.

Some features are not yet implemented. The star and save buttons on the image viewer do not yet do anything. They also don't work in the legacy app so that's not too much of an issue.

Also, one final note, I am not a Java nor Kotlin dev. It's all new territory.